use float field for NoNPCDamageBelowHealthPct

+++ sql/updates/world/master/9999_99_99_99_world.sql
@@ -2,7 +2,7 @@
 DROP TABLE IF EXISTS `creature_template_sparring`;
 CREATE TABLE `creature_template_sparring` (
   `Entry` int unsigned NOT NULL,
-  `NoNPCDamageBelowHealthPct` tinyint unsigned NOT NULL,
+  `NoNPCDamageBelowHealthPct` float NOT NULL,
   PRIMARY KEY (`Entry`,`NoNPCDamageBelowHealthPct`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
